build: fix example config to work out-the-box
BODS_AWS_REGION always needs to be set, even if running
`bin/transform_local`.

BODS_AWS_ACCESS_KEY_ID and BODS_AWS_SECRET_ACCESS_KEY need to be set to
a non-empty value, but do not have to be valid.

BODS_STREAM and PSC_STREAM must be unset, not simply empty, in order to
not error.

With this commit, documents can be indexed using register-v2
Elasticsearch, with minimal setup. To do so:

register-v2:

- copy `.env.example` to `.env`
- `docker compose up`

register-transformer-psc:

- copy `.env.example` to `.env`
- `docker compose up`
- `docker compose exec transformer-psc setup_indexes_companies`
- `docker compose exec transformer-psc setup_indexes`
- `docker compose exec transformer-psc transform_local statements/all4.jsonl`