build: generate RSpec binstub
Otherwise, `rspec` command can't be used without `bundle exec` prefix.

Alternatively, `bundle exec` prefix could be added to `bin/test` script
when calling RSpec. I'm not certain which is the best approach, here,
but it is common to be able to run RSpec without needing a prefix.

With this commit tests pass within Docker:

    docker compose exec transformer-psc test