[#189] upgrade common, sources-bods for migrations structure
Replace the need to have a separate bin script for each migration with a
generic script which will dispatch to the correct migration.

`migrate-lei` script is removed, since it's been run already. If it's
needed again at some point, `MigrateLEI` approach should be ported to the
new migrations approach in Sources BODS.